+++ backend/azure/location_test.go
@@ -36,7 +36,7 @@ func (s *LocationTestSuite) TestString() {
 
 func (s *LocationTestSuite) TestList() {
 	client := mocks.NewClient(s.T())
-	client.EXPECT().List(mock.Anything).Return([]string{"file1.txt", "file2.txt"}, nil)
+	client.EXPECT().List(mock.Anything, mock.Anything).Return([]string{"file1.txt", "file2.txt"}, nil)
 	fs := NewFileSystem(WithClient(client))
 	l, _ := fs.NewLocation("test-container", "/some/folder/")
 	listing, err := l.List()
@@ -46,7 +46,7 @@ func (s *LocationTestSuite) TestList() {
 
 func (s *LocationTestSuite) TestListByPrefix() {
 	client := mocks.NewClient(s.T())
-	client.EXPECT().List(mock.Anything).Return([]string{"file1.txt", "file2.txt", "foo.txt"}, nil)
+	client.EXPECT().List(mock.Anything, mock.Anything).Return([]string{"file1.txt", "file2.txt", "foo.txt"}, nil)
 	fs := NewFileSystem(WithClient(client))
 	l, _ := fs.NewLocation("test-container", "/some/folder/")
 	listing, err := l.ListByPrefix("file")
@@ -58,7 +58,7 @@ func (s *LocationTestSuite) TestListByPrefix() {
 
 func (s *LocationTestSuite) TestListByRegex() {
 	client := mocks.NewClient(s.T())
-	client.EXPECT().List(mock.Anything).Return([]string{"file1.txt", "file2.txt", "foo.txt"}, nil)
+	client.EXPECT().List(mock.Anything, mock.Anything).Return([]string{"file1.txt", "file2.txt", "foo.txt"}, nil)
 	fs := NewFileSystem(WithClient(client))
 	l, _ := fs.NewLocation("test-container", "/some/folder/")
 	regex := regexp.MustCompile("file")
@@ -103,7 +103,7 @@ func (s *LocationTestSuite) TestPath() {
 
 func (s *LocationTestSuite) TestExists() {
 	client := mocks.NewClient(s.T())
-	client.EXPECT().Properties("test-container", "").Return(&BlobProperties{}, nil)
+	client.EXPECT().Properties(mock.Anything, "test-container", "").Return(&BlobProperties{}, nil)
 	fs := NewFileSystem(WithClient(client))
 	l, _ := fs.NewLocation("test-container", "/some/folder/")
 	exists, err := l.Exists()
@@ -113,7 +113,7 @@ func (s *LocationTestSuite) TestExists() {
 
 func (s *LocationTestSuite) TestExists_NonExistentFile() {
 	client := mocks.NewClient(s.T())
-	client.EXPECT().Properties("test-container", "").Return(nil, errors.New("no such file"))
+	client.EXPECT().Properties(mock.Anything, "test-container", "").Return(nil, errors.New("no such file"))
 	fs := NewFileSystem(WithClient(client))
 	l, _ := fs.NewLocation("test-container", "/some/folder/")
 	exists, err := l.Exists()
@@ -258,7 +258,7 @@ func (s *LocationTestSuite) TestNewFile_NilReceiver() {
 
 func (s *LocationTestSuite) TestDeleteFile() {
 	client := mocks.NewClient(s.T())
-	client.EXPECT().Delete(mock.Anything).Return(nil)
+	client.EXPECT().Delete(mock.Anything, mock.Anything).Return(nil)
 	fs := NewFileSystem(WithClient(client))
 	l, _ := fs.NewLocation("test-container", "/some/folder/")
 	s.Require().NoError(l.DeleteFile("clever_file.txt"), "the file exists so we do not expect an error")
@@ -266,7 +266,7 @@ func (s *LocationTestSuite) TestDeleteFile() {
 
 func (s *LocationTestSuite) TestDeleteFile_DoesNotExist() {
 	client := mocks.NewClient(s.T())
-	client.EXPECT().Delete(mock.Anything).Return(errors.New("no such file"))
+	client.EXPECT().Delete(mock.Anything, mock.Anything).Return(errors.New("no such file"))
 	fs := NewFileSystem(WithClient(client))
 	l, _ := fs.NewLocation("test-container", "/some/folder/")
 	s.Require().Error(l.DeleteFile("nosuchfile.txt"), "the file does not exist so we expect an error")