+++ backend/ftp/newFileSystemOption_test.go
@@ -1,22 +1,21 @@
-package gs
+package ftp
 
 import (
 	"context"
-	"reflect"
 	"testing"
 
-	"cloud.google.com/go/storage"
+	"github.com/jlaffaye/ftp"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestWithClient(t *testing.T) {
-	client := &storage.Client{}
+	client := &ftp.ServerConn{}
 	fs := &FileSystem{}
 
 	opt := WithClient(client)
 	opt.Apply(fs)
 
-	assert.Equal(t, client, fs.client, "Client should be set correctly")
+	assert.Equal(t, client, fs.ftpclient, "Client should be set correctly")
 }
 
 func TestWithOptions(t *testing.T) {
@@ -37,15 +36,4 @@ func TestWithContext(t *testing.T) {
 	opt.Apply(fs)
 
 	assert.Equal(t, ctx, fs.ctx, "Context should be set correctly")
-}
-
-func TestWithRetryer(t *testing.T) {
-	retryer := func(wrapped func() error) error { return wrapped() }
-	fs := &FileSystem{}
-
-	opt := WithRetryer(retryer)
-	opt.Apply(fs)
-
-	assert.NotNil(t, fs.retryer, "Retryer should not be nil")
-	assert.Equal(t, reflect.ValueOf(retryer).Pointer(), reflect.ValueOf(fs.retryer).Pointer(), "Retryer should be set correctly")
 }